# backend-lecture-code/day17-microservices/microservices/gateway/openapi.yaml
openapi: 3.0.3
info:
  title: Day 17 Store — Microservices (via the API Gateway)
  version: 1.0.0
  description: |
    The same store as the monolith, split into eight services behind one gateway. The public
    paths are identical to the monolith's; only the port changed (4100 → 8080).

    **Log in first:** call `POST /users/login` with `{"email": "alice@shop.com"}`, copy the
    `token`, then click **Authorize** and paste it. The gateway checks the token once and passes
    the caller's id to the service in an `x-user-id` header. Services never see the token.

    Any routed request can also fail at the gateway itself: **502** when the service is down,
    **504** when it takes longer than 10 seconds.

    The **internal** section at the bottom lists service-to-service endpoints. The gateway has no
    route for them; they're documented with each service's own address for reference.

    Seed data to know: the 85" TV (`SKU-2001`, $5,999) is over the $5,000 limit, so paying for
    it is always declined. The Coffee Mug (`SKU-3001`) has only 1 in stock.
servers:
  - url: http://localhost:8080
    description: API gateway
...
paths:
...
  /users:
    post:
      tags: [users]
      summary: Sign up
      security: [] # public, no token needed
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/NewUser" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "409": { $ref: "#/components/responses/Conflict" }
        "502": { $ref: "#/components/responses/ServiceDown" }
  /users/login:
    post:
      tags: [users]
      summary: Log in (demo — email only, no password)
      security: [] # public, no token needed
      description: The users service signs the JWT; the gateway verifies it on later requests.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginRequest" }
      responses:
        "200":
          description: A JWT, valid for one hour
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LoginResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "502": { $ref: "#/components/responses/ServiceDown" }
...
  /products:
    get:
      tags: [catalog]
      summary: List products
      security: [] # public, no token needed
      description: |
        Round-robins across every catalog instance in `CATALOG_URLS`. Start a second instance
        and watch the gateway log alternate between them.
      parameters:
        - name: category
          in: query
          required: false
          schema: { type: string, example: electronics }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Product" }
        "502": { $ref: "#/components/responses/ServiceDown" }
    post:
      tags: [catalog]
      summary: Add a product
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Product" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Product" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "409": { $ref: "#/components/responses/Conflict" }
        "502": { $ref: "#/components/responses/ServiceDown" }
  /products/{sku}:
    get:
      tags: [catalog]
      summary: One product
      security: [] # public, no token needed
      parameters: [{ $ref: "#/components/parameters/Sku" }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Product" }
        "404": { $ref: "#/components/responses/NotFound" }
        "502": { $ref: "#/components/responses/ServiceDown" }
  /inventory:
    get:
      tags: [inventory]
      summary: Stock for every SKU
      security: [] # public, no token needed
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Stock" }
        "502": { $ref: "#/components/responses/ServiceDown" }
  /inventory/{sku}:
    get:
      tags: [inventory]
      summary: Stock for one SKU
      security: [] # public, no token needed
      parameters: [{ $ref: "#/components/parameters/Sku" }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Stock" }
        "404": { $ref: "#/components/responses/NotFound" }
        "502": { $ref: "#/components/responses/ServiceDown" }
...
  /internal/inventory/reservations:
    post:
      tags: [internal]
      summary: Reserve every item or none (called by orders)
      security: [] # public, no token needed
      servers: [{ url: http://localhost:4103, description: inventory service }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [items]
              properties:
                items:
                  type: array
                  items: { $ref: "#/components/schemas/CartItemInput" }
      responses:
        "201":
          description: Reserved
          content:
            application/json:
              schema:
                type: object
                properties:
                  reservationId: { type: integer, example: 1 }
        "400": { $ref: "#/components/responses/BadRequest" }
        "409": { $ref: "#/components/responses/Conflict" }
  /internal/inventory/reservations/{id}:
    delete:
      tags: [internal]
      summary: Release a reservation (the compensating action)
      security: [] # public, no token needed
      description: Puts the stock back. Releasing the same reservation twice does nothing.
      servers: [{ url: http://localhost:4103, description: inventory service }]
      parameters: [{ $ref: "#/components/parameters/Id" }]
      responses:
        "204": { description: Released }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Sku:
      name: sku
      in: path
      required: true
      schema: { type: string, example: SKU-1001 }
    Id:
      name: id
      in: path
      required: true
      schema: { type: integer, example: 1 }
  responses:
    BadRequest:
      description: Invalid input
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Missing, invalid, or expired token (checked at the gateway), or unknown email on login
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Conflict:
      description: Already exists, or not enough stock
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    ServiceDown:
      description: The service behind this route is unreachable (from the gateway itself)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/UpstreamError" }
  schemas:
    Error:
      type: object
      properties:
        error: { type: string, example: "unknown sku NOPE" }
    UpstreamError:
      type: object
      properties:
        error: { type: string, example: "payments timed out after 3000ms" }
        service: { type: string, example: payments }
...
    User:
      type: object
      properties:
        id: { type: integer, example: 1 }
        email: { type: string, example: alice@shop.com }
        name: { type: string, example: Alice }
    NewUser:
      type: object
      required: [email, name]
      properties:
        email: { type: string, example: dave@shop.com }
        name: { type: string, example: Dave }
    LoginRequest:
      type: object
      required: [email]
      properties:
        email: { type: string, example: alice@shop.com }
    LoginResponse:
      type: object
      properties:
        token: { type: string }
        user: { $ref: "#/components/schemas/User" }
    Product:
      type: object
      required: [sku, name, price, category]
      properties:
        sku: { type: string, example: SKU-2001 }
        name: { type: string, example: 85" 8K TV }
        price: { type: number, example: 5999 }
        category: { type: string, example: electronics }
        attributes:
          type: object
          description: Different for every kind of product (stored in MongoDB, so no fixed schema).
          additionalProperties: true
          example: { resolution: 8K, screenInches: 85 }
    Stock:
      type: object
      properties:
        sku: { type: string, example: SKU-1001 }
        onHand: { type: integer, example: 25 }
    CartItemInput:
      type: object
      required: [sku]
      properties:
        sku: { type: string, example: SKU-1001 }
        quantity: { type: integer, minimum: 1, default: 1, example: 2 }
...